pyproject.toml
[project]
name = "celine-rec-registry"
version = "0.1.0"
description = "CELINE REC Registry microservice (read-only JSON-LD API)"
readme = "README.md"
requires-python = ">=3.11"
authors = [{name = "CELINE"}]
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "sqlalchemy>=2.0",
  "asyncpg>=0.29",
  "psycopg[binary]>=3.2",
  "alembic>=1.13",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "typer>=0.12",
  "pyyaml>=6.0",
  "python-dotenv>=1.0",
]

[project.scripts]
celine-rec-registry = "celine.rec_registry.cli.main:app"

[tool.uv]
package = true

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["celine*"]

[tool.setuptools.package-data]
"celine.rec_registry.resources" = ["*.jsonld", "*.ttl", "*.md"]

[dependency-groups]
dev = [
    "dumpster-cli>=1.3.2",
]
